firebot: correction to kill option for mac

+++ Firebot/run_firebot.sh
@@ -145,6 +145,11 @@ if [ $notfound -eq 1 ] ; then
   QUEUE=none
 fi
 
+platform="linux"
+if [ "`uname`" == "Darwin" ] ; then
+  platform="osx"
+fi
+
 USEINSTALL=
 BRANCH=master
 botscript=firebot.sh
@@ -262,8 +267,10 @@ shift $(($OPTIND-1))
 if [ "$KILL_FIREBOT" == "1" ]; then
   if [ -e $firebot_pid ] ; then
     PID=`head -1 $firebot_pid`
-    echo killing process invoked by firebot
-    kill -9 $(LIST_DESCENDANTS $PID)
+    if [ "$platform" == "linux" ]; then
+      echo killing process invoked by firebot
+      kill -9 $(LIST_DESCENDANTS $PID)
+    fi
     echo "killing firebot (PID=$PID)"
     kill -9 $PID
     if [ "$QUEUE" != "none" ]; then
